move monitors/previews to their own files
